network: stabilize host & DC state on change cluster
Moving a host to another cluster (aka 'change cluster') should preserve
the state of the host and the data center. Change cluster starts with
deactivating the host and ends with reactivating it.

There are two situations to be considered when changing the cluster of a
host:

1. When there are no storage domains in the data center (or they are in
'Maintenance') the data center remains in an 'Uninitialized' state
throughout the process of changing cluster, so waiting for the data
center to be UP blocks indefinitely. Therefore the unconditional wait
for data center UP is being removed in this patch.

2. When a storage domain exists and is up in the data center, one of the
hosts must be designated as SPM. When this host is deactivated during
change cluster, it loses its SPM status and as a consequence the data
center becomes 'Non Responsive'.
The data center will be UP again only once a host regains SPM status. It
may be designated to the same host once it is reactivated if it is the
only active host in the data center. Or it may be designated to another
host immediately after the deactivation of the first host.

So to address the fluctuation in the data center status in both of the
above situations it is only required to wait for the data center to
regain UP status if the host which is changing cluster was an SPM before
the change cluster started.
If the host which is changing cluster was not an SPM then either:
- there are no SPMs because there are no active storage domains
- another host is the SPM, and will remain so, so the data center will
  remain UP the whole duration of the change cluster.

Moving the check for data center UP status from toggle_cluster to
change_cluster also ensures that the stabilization is awaited for
immediately after each destabilization for current and future uses.

Change-Id: Ia060bda284bbfd41ae0c33978060dc191f3cf920