basic: Align basic suite master with pytest naming
All test modules should start with test_ that way
pytest is able to recognize them and pick them
automatically.

This allows to pass only folder to pytest instead of
individual modules in lagofy.sh and control.sh.

At the same time we need to create local copy  of few
test modules for performance suite because it is still
not completely pytest compatible.

Change-Id: I0cd98daab3a3be1aa90a3c53c5f4a6a9b39f1547
Signed-off-by: Ales Musil <amusil@redhat.com>