he: Generate ansible inventory
This patch calls the 'env_dump_ansible_hosts' function defined
in 'run_suite.sh' during deployment of he suite.

This function generates an ansible inventory in
'deployment-he-basic-suite-master/hosts' file, which is used by all
'ansible_*' fixtures to communicate with the VMs.

We also need to add the 'python2-ansible-runner' package as a dependency
so that ansible modules can be executed successfully.

Change-Id: I058ff418bbe966b61fe98516e7310659ae1d1c52
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>