he: Add initial support for prebuilt OST images
This patch prepare HE suite to use prebuilt OST images.

We don't have an image with HE appliance installed yet, so we're
using 'ost-images-*-host-installed' as a base. For the storage VM
we're talking the 'ost-images-*-upgrade' which contains a bare distro
+ upgrades.

Prebuilt OST images have their root SSH key baked in, so we skip
the usual key injection done by lago. That's why we need a hook
for HE deployment to copy the baked-in SSH key to the HE VM.

The directory for the hooks seemed to change, so the path has been
updated for the existing hooks.

Change-Id: I616f8ff008d193bb4fa9efcbca35117275ddd0d9
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>