ost_utils: Move 'versioning' module to 'ost_utils'
With [1] we unfortunately broke a lot of suites.

The problem is that basic suite's 'test_utils' directory is
symlinked all over the place and other suites are using functions like
'guest_os_image_name'. We need to get rid of this symlinking hell.

For now we're:
- moving 'versioning' module to 'ost_utils', so it's clear
  that the thing is shared
- restoring removed functions to fix broken suites
- removing all remaining 'versioning' modules accross different suites
- reimplementing 'virt' fixtures in terms of 'ost_utils.versioning'
  functions

'test_utils' must die.

[1] https://gerrit.ovirt.org/#/c/112196/

Change-Id: I660bf75567be0a06ea43d253d5978337113278ad
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>