network: remove lago env dependency from fqdn fixture
Acquiring the details of engine and hosts and ssh-ing into them is now
accomplished with ansible_facts and sshlib functionality, so lagosdk API
is no longer needed in fqdn.py fixtures.

Although these fixtures no longer call test-scenarios/conftest.py#env
it is nevertheless guaranteed that lago deployment has already happened
when they are invoked since they are called for the first time in
test-scenarios/test_ipv6_support.
Therefore it is guaranteed that engine and the hosts are up and
responsive to ssh connections.

Change-Id: Ia81f5ab5889006d8ed4d217a63eae37b8f865b86
Signed-off-by: Eitan Raviv <eraviv@redhat.com>