network: host de/activate - add retry
Activating a host may fail on first attempt. Retrying until the host is
up is introduced in this patch because it is more likely to succeed. But
an attempt to activate a host which is already UP is denied with an
exception. This may occur with a retry mechanism because the host can
reach status UP while the retry mechanism is asleep. This situation is
handled by expanding syncutil#sync to check for success also after an
exception is thrown.

The same reasoning has been applied to deactivating a host.

Change-Id: I5e450b7ed74d0ff1ac5a669d7fb51cd5cd178bd3
Signed-off-by: Eitan Raviv <eraviv@redhat.com>