UI tests: Improve tests for the left nav menu
The left navigation menu was not tested properly -
the test just clicked through the menu and created
screenshots (no assertions)

This patch removes the special test method for menu.
Instead, there are new test methods that test
the concrete entities.

The typical minimal scenario is to click the entity in the
left menu, check if the entity is in the table, select it
and check the enablement of the action buttons.

Change-Id: If7c10789123bd7dee65a275bec390f107b154324
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>