test_utils/host_status_utils: Make dc_name parametric
Some, but not all, of the functions in this module, already had a
parameter dc_name. In HE suites, we do not create another DC (and
cluster), but use Default, because it's already created as part of
deployment. Make all functions there accept dc_name as a parameter,
so that HE suites can also use it.

Change-Id: I40fb6d6029028c073bc153820a91ccc3dd5a33a0
Signed-off-by: Yedidyah Bar David <didi@redhat.com>