basic: Workaround for hostnames issue
Hostnames of the VMs have a network iface name appended
in the middle of suite run, so they change i.e. to:

 lago-basic-suite-master-host-0-eth1

This is tracked and better documented in the bug [1].

Until we get a proper fix let's set the hostnames at the beginning
of the suite, so other tests can rely on it.

[1] https://issues.redhat.com/browse/RHV-39904

Change-Id: I1b3fe625647b13a38bfdecd45c637645930181a8
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>