network: fix formatting errors
Fix formatting errors caused by [1].
f-strings recently introduced in 'ost_utils' parts of the codebase
which seem to create conflict on flake8's logic.
upgrade the flake8 in order to solve the issue.
on this occasion upgrade pylint version to align it with the one used
globally in tox.ini.
this will ease the transition of linters from network suite to the
global ones.
[1] https://gitlab.com/pycqa/flake8/-/issues/240

Change-Id: Ife5affce76ddc525bffd1777f2ab00e13f657a34
Signed-off-by: hbraha <hbraha@redhat.com>