ost_utils: ansible: Use in-memory layer of caching for facts
'ansible-runner' keeps a cache of information on a VM as a set
of files in a private folder. That includes facts obtained with
'gather_facts' module and provided with 'ost_utils.ansible.facts.*'
helpers.

Until now, when someone queried a fact about a VM, we went through
the cached files to find appropriate response every time. This is being
improved in current patch - we're adding a second layer of in-memory
cache for queried facts, so if someone asks for i.e. "ansible_hostname"
for the second time, the response will be quickly served from a python
dictionary.

Change-Id: Ic59bd2cc6a489037ed93b673c5b1271f4168eff4
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>