flake8: Rearrange list of ignores
While keeping the list of flake8's ignores sorted alphabetically
is nice, it makes more sense to sort them by the frequency of their
occurence. That way we can start fixing them starting with the ones
that occur the least amount of times.

The occurence count for each of these ATM of writing is:

  1 E127
  1 E222
  1 E301
  1 E721
  2 E211
  2 E241
  2 E262
  2 F402
  2 W391
  3 E266
  3 W503
  4 E131
  5 E126
  5 E203
  5 E712
  6 E123
  7 E111
  7 E251
  7 F405
  7 F841
  7 W191
  8 E128
  8 E265
  9 E121
  9 E226
  9 F821
 10 E303
 11 E225
 12 E722
 13 E101
 13 E202
 14 E122
 21 E231
 25 F811
 30 F403
 43 E302
 50 E261
112 F401
236 E501

Change-Id: I7bc6b7fec46368a7532477d8b7a14b632e6092f9
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>