ost_utils: backend: Add 'artifacts' method
This patch adds a new abstract 'artifacts' method
to the 'BaseBackend' class and a concrete implementation
for the only backend we have at the moment - 'LagoBackend'.

The method returns a dictionary which maps hostnames to a list
of files, i.e.:

 {
     'lago-basic-suite-master-engine': [
         '/var/log',
         '/var/cache/ovirt-engine',
     ],
     'lago-basic-suite-master-host-0': [
         '/etc/resolv.conf',
         '/var/log',
     ],
     'lago-basic-suite-master-host-1': [
         '/etc/resolv.conf',
         '/var/log',
     ]
 }

The lago implementation works by parsing and extracting information
from 'LagoInitFile'.

This is another step in gaining lago independence. While conceptually
a backend does not determine the list of artifacts we want to gather
(meaning if we had an i.e. oVirt backend, we'd still want to collect
the same files), this piece of information is a good candidate
for being proxied by the 'backend' module (i.e. from something we could
call in the future an 'OSTInitFile' instead of 'LagoInitFile').

Change-Id: I4d41b9d9d7d57ee7c9669f2fd56f03715270c503
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>