ost_utils: Move test_utils.selenium to ost_utils.selenium
Pylint in the past had some issues with symlinking, that we
thought we solved [1]. However, this solution does not work,
since it hides selenium package from pylint. To overcome this
we are moving all the selenium stuff to ost_utils, fixing
some of the more convoluted imports and replacing symlinks
with interfaces.

[1] https://gerrit.ovirt.org/c/ovirt-system-tests/+/113030

Change-Id: Ic8ee69aada76b1042a498c57c7a37cd11133e146
Signed-off-by: Andrej Cernek <acernek@redhat.com>