performance: Remove suite
This patch removes performance suite from the repository.

Galit's recent work around it [1] made the code much more modern
and on par with other suites. The gist of the performance suite
however, is testing the engine with a vast number of hosts.
We have a project for this [2], but it's not really maintained
by anyone at this moment. Unfortunately we don't have neither
the manpower to pick it up again nor the HW resources to use
large number of real hosts for testing. For these reasons
let's remove it from the repository for now and focus on the most
important suites.

[1] https://gerrit.ovirt.org/#/q/status:merged+project:ovirt-system-tests+branch:master+topic:perf_update
[2] https://gerrit.ovirt.org/admin/repos/ovirt-vdsmfake

Change-Id: I4133e591bd1d5e8cfd9e82bff2c208571724c312
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>