vagrant: Remove suite
This patch removes the basic suite based on Vagrant
from the repository.

We use Lago as a VM orchestrator for all the suites. There's
no one really to maintain it and it failed to gain traction
as the open source choice for it's purpose. Therefore, we were
trying to make a switch to Vagrant.

With the chosen path of implementation, this effort turned out
to be too hard - it was a constant race to follow all the changes
happening around in OST project.

Currently there is a different effort that tries making OST
backend-independent, which may lead to a Vagrant switch in the future.
The implementation however relies on abstracting-away
the backend instead of reimplementing parts of Lago API
in terms of Vagrant. It also doesn't require an explicit fork
of the basic suite to be functional. For these reasons it's better
to have the fork removed and focus on improving 'ost_utils.backend'
module.

Change-Id: I18acef8d2242bf8d1724a99ef536f1b9beb0e09f
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>