he: Remove ipv6, iscsi and role-remote variants
This patch removes 3 variants of the HE suite.

While it's desirable to have the possibility to configure
suites to use i.e. ipv6 only or iscsi storage domains, we should
not implement this with the means of separate directories
and symlinks. It makes maintaining the code very hard.
On top of that, we don't really have the HW resources to i.e.
run 4 variants of HE suite at the same time.

Let's focus on the main 'he-basic-suite-master' for now and
work on providing configurability and modularization in OST
with different means.

Change-Id: I6eb458ee94327c63f2d113005e979ee096fac56d
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>