ansible: Don't depend on other fixtures in 'ansible_by_hostname'
With 'ansible_by_hostname' fixture being dependent on other
fixtures, suites that use 1 host (like the new node upgrade suite
that's still WIP) break here [1].

Since the recent refactoring the fixture's definition becomes trivial.
We can replace the fixture with direct calls in the future, but we
already depend on it in multiple places and it's not that important.

[1] https://github.com/oVirt/ovirt-system-tests/blob/3b58a65b8adb7adb33f27a073792e951594f677a/ost_utils/ost_utils/ansible/patterns.py#L37

Change-Id: I3b6e5c7eeaec9e6644818ebac4ba075fc37ac94b
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>