pytest: Introduce 'backend' fixture
Since introducing the 'ost_utils.backend' module recently
we had calls like 'backend.default_backend()' scattered around
the codebase. This approach however is fragile in terms
of potential issues with refactoring in the future.

With this patch we introduce a level of indirection by wrapping
the call in a fixture that's used by different tests.

Change-Id: Id9ce7316b6938bf3d5b1bb74faf3aba3f1da7a13
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>