basic: Reboot only one of the hosts
With recent patch to engine [1] all deployed hosts are automatically
rebooted. This behavior is beneficial from the customers' point of view, since
it mitigates many potential issues, but in OST we want to move fast.

With this patch we'll be rebooting only one of the hosts during
deployment. This will allow us to continue with some actions faster
(since some of the tests require only 1 host to be up) and test
the reboot process at the same time.

Additionally, since OST's VMs boot fast, we shorten the reboot timeout
in the engine config to 120s.

[1] https://gerrit.ovirt.org/#/c/ovirt-engine/+/113332/

Change-Id: Id0a589f86e7014ccaecabe16945ae7fcd9054a3d
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>