basic: Make 'test_set_hostnames' more dynamic
We have the 'test_set_hostnames' in basic suite to mitigate
an issue described here [1]. We're close, but we still don't have
a better way to handle this.

The problem with the test is that it assumes there are always at least
2 hosts available, which is not true for i.e. the new terraform suite
that's WIP. Let's have a quick fix for this issue for now.

[1] https://issues.redhat.com/browse/RHV-39904

Change-Id: I7fcdb95f8fed99b0eab506e0a6aae1a9da5c6870
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>