Use pytest 6.2.2
We intend to adapt some test modules so that they are usable for more
than one suite. We intend to then symlink them between suites.

pytest 4.6.9 behaves differently from other versions (including 6.2.2)
in this regard: It uses conftest.py from the directory of the symlink
target, not of the source. This means that it's impossible to cause
different suites to have different behavior for a particular test module
by only overriding some fixtures in conftest.py.

We originally picked 4.6.9 because it was the last version supporting
python2. Since this isn't needed anymore, there should be no problem
upgrading. Might as well upgrade to current latest, which is 6.2.2.

Change-Id: I19bd39f6a237693628d184780b9fb4c4e8f3e02b
Signed-off-by: Yedidyah Bar David <didi@redhat.com>