ansible: switch to collection instead of sdk
This patch adds ansible_setup to the ost_utils which is used to
create playbook from the template, then copy it to the engine VM
and execute it ansible-runner.
We copy it so we use the collection with which was the engine
installed.
For the roles run we don't gather any infromation about the run. It stops
the test if the role fails. But for the modules, we get an ansible response.

Change-Id: I087399ed03495eb8f355f929676b5c59d7ea56f5
Signed-off-by: Martin Necas <mnecas@redhat.com>