basic: ui: Use 'podman-remote' in mock envs
With recent update of 'podman' we are finally able to use
podman for UI tests in mock-based CI.

Until now, when running 'podman pod create' there was no way
to define the network mode used for the pod. For rootless
containers it was 'slirp4netns' and for rootful containers
it was 'bridge'. Given that, and the fact that CI pipeline
is running as 'root', we were locked in to the 'bridge' mode,
which prevented the Selenium grid from being able to access
the engine's webadmin UI.

In the latest version of 'podman' a '--network=mode' argument
has been added to 'pod create', so we can finally run the UI
tests in mock envs.

This patch requires a change in the 'global_setup' script that
will expose a podman socket on CI agents with:

 podman system service --time=0 unix:///var/run/podman/podman.sock

Change-Id: Ie52333ccf531cf27b58a8faf0c33825b2178a1e9
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>