basic: Add 'python3-requests' dependency
We depend on the 'requests' module in one of basic suite's
test modules [1], but we never listed is as an explicit
dependency. Recently, some CI jobs started failing on missing
'requests' module. It's possible we always had it cause it's
a mock dependency and the new envs do not use mock. Nevertheless,
if we use it, we should mention it explicitly in the '*packages*'
files.

[1] https://github.com/oVirt/ovirt-system-tests/blob/adf5c83f62f0888a9eb8a55fa9faff5e20a71d83/basic-suite-master/test-scenarios/test_098_ovirt_provider_ovn.py#L26

Change-Id: I4ee29ebcbee464ef56baa9009c646466e591a4bd
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>