he: Remove 'ipv6_utils' usage
HE suite used 'ipv6_utils' module to test the API with
an ipv6 connection. The implementation used some deprecated
stuff from lago OST plugin and we want to get rid of that
dependency. We need a more global approach for handling ipv6,
so in this patch we drop the remaining usage of 'ipv6_utils'
module and the module itself.

Change-Id: Ie2b6af9c80ff32b36bcc816d29068bd4a71cd380
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>