compat: Remove suite
Compat suite was there as a variant of basic suite to test
compatibility with older cluster levels. ATM it seems
that there's not much of a difference between this one and
the basic suite. It's not worth having a completely separate
suite for these kind of things - it should be a parameter
to the basic suite. Let's remove it and work on parametrization
instead.

Change-Id: I662faa5b68db30a52199e0348012e8751beae2e5
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>