upgrade-from-release: Remove suite and related files
With 4.4 there's no way we can test upgrades like we did
before since we require a full reinstallation on el8.
It doesn't really make sense to keep the suite in the codebase.

We should work on a suite that would allow testing migration
from el7 nodes to el8 nodes on el8 engine, but that's a different
subject.

Change-Id: Ied03864bc6d6bb9068785e0b4d2268b9513bbe6c
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>