flake8, network: Use tox flake8 run on network suite
Our in intention in having 2 separate runs of flake8, 1 for network
and 1 for the rest, was that network suite is in better state and we
needed to disable some checks for other suites. This approach however
backfired, as the network flake8 used tox.ini config anyway, including
the ignores.

This patch separates the ignores per folder, so that each suite gets
the best linting coverage it can. This way we can also use just one
flake8 run and avoid cluttering of the enviroment. To do all of this
we need flake8>=3.7.0, so we can as well use the newest one.

Change-Id: I0bc6d5387e41177cafbc3fe65de033abc46f5c05
Signed-off-by: Andrej Cernek <acernek@redhat.com>