ost_utils: Function for running deployment scripts
This patch introduces a new 'deployment_utils.scripts'
module with a 'run_scripts' function. It does the same
thing as 'lago deploy' we use now - it runs the deployment
scripts listed in 'LagoInitFile' on the VMs.

In next patches we'll move away from using 'lago deploy'
in favour of this function.

Change-Id: I9fdc1b9ced2c9b653d0ed63e889bffbda6a88709
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>