network: adaptable machine facts
ost_utils.ansible.ansible_facts retrieves information from the remote
machine whenever its get attribute is called. When an ansible service is
present and fetches the facts MachineFacts caches the ansible response
as a simple collection of machine access info.
When an ansible service is not present in an independent environment
MachineFacts can be easily instantiated with access info for the engine
and hosts in the environment, enabling fast and easy setup for testing
and debugging.

MachineFacts therefore serves as an adapter that encapsulates dependency
upon the ost_utils-ansible stack from the network suite for multiple
debugging and testing environments.

Change-Id: I22af0e263b4619135c2ecfb858ff6c944aafb891
Signed-off-by: Eitan Raviv <eraviv@redhat.com>