basic: xfail live storage migration test
'test_live_storage_migration' fails randomly. After investigation
it turned out the problem lies somewhere deep in libvirt. A bug has
been filed for this.

Since the issue is known and tracked let's try xfailing the test.
It's possible that when the problem occurs we'll fail on other tests
due to repercussions of the libvirt issue, but we'll have to find
that out and revert the xfail in that case.

Change-Id: I7760d67faa456303d7428aef6bd622852e0bbb4e
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>