network: README - how to debug run on independent deployment
Developers of both engine\vdsm and OST need the following capabilities:

* test their code changes on engine\vdsm against network suite during
  development with zero setup time and zero wait-for-results time.
* debug the suite code in a visual debugger and use break points to stop
  the execution at a desired point.

Network suite was designed with these goals in mind and it is possible
to run it against any independent deployment of engine and hosts to
accomplish the above goals.
A user need only supply or modify very few details to achieve a
functioning independent run. The README details these adjustments for
the time of writing.

Change-Id: Ied6f372783d8874575c04d2df82267ee57643f60
Signed-off-by: Eitan Raviv <eraviv@redhat.com>