selenium: podman: Find free ports in grid setup retries
In CI environment from time to time we trigger two or more
basic suite runs at the same time. This may cause races in Selenium
grid setup - both grids will try to use the same port and one
of them will fail.

We already have a mechanism for retrying Selenium grid setup
to overcome random instabilities in podman. Since the port
number used for the grid is actually an implementation detail
exposed to the clients only in the returned URL, let's move
the code that searches for a free port to the retry logic.
That should mitigate these issues.

Change-Id: I4e1a06b1f0272def2eb58979ce070073ad0d5208
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>