ansible: Make returning results for multiple hosts working
Until now this was possible:

 result = ansible_host0.ping(data='panda')
 assert result['ping'] == 'panda'

Using this on a mapper that covered more than one VM didn't
work though:

 result = ansible_hosts.ping(data='panda')
 assert result['ping'] == ???

the result of this operation would come from one of the hosts,
but we'd never knew which one.

This patch changes that - when the mapper covers more than one
VM, the results will be groupped in a dictionary that maps
hostnames to results:

 result = ansible_hosts.ping(data='panda')
 assert result['lago-basic-suite-master-host-0']['ping'] == 'panda'
 assert result['lago-basic-suite-master-host-1']['ping'] == 'panda'

The behavior in case where the mapper covers only one host doesn't
change.

Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>
Change-Id: Icae15217cceac761937a803868f631fb944b2bce