deploy: Always show list of used oVirt packages
Until now we've only shown the list of oVirt-related packages
in case someone used custom repositories in the run (and the ones
that are installed in ost-images if you're using lagofy). The list
was also also limited only to the packges that came from these
custom repos.

With this patch we're always logging a summary of oVirt-related
packages used in the VMs, sorted alphabetically.

Change-Id: I2c3cd876815523c798b1a138695a3831d11d5f3a
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>