package_mgmt: Don't use 'hostnames' in 'check_used_packages'
Since now we always report the list of used oVirt packages [1]
and we can retrieve results from ansible calls that cover more than
one VM [2] we don't really need to use the 'hostnames' fixture anymore.
The switch to 'ansible_all' is desired from HE's suite point of view
and the advantage of this move is taken in subsequent patches.

[1] https://gerrit.ovirt.org/114856
[2] https://gerrit.ovirt.org/114858

Change-Id: Ic6a1fa3de21bd6907faea3c00dcc62adec2031c6
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>