selenium: Don't define network backend in rootless
In u/s we run the tests in mock as the root user. The default network
backend used by podman is different for rootless/rootfull modes,
so we needed to change it in [1]. This however broke d/s pipelines
where we run as non-root. Let's omit the '--network=slirp4netns'
parameter if we're running rootless.

[1] https://gerrit.ovirt.org/#/c/ovirt-system-tests/+/113715/6/ost_utils/ost_utils/selenium/grid/podman.py

Change-Id: Id81a07d6a7fbefe6ff1d1992bac497b382c1fac1
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>