deployment: Use 'ansible_vms_to_deploy' instead of 'ansible_all'
This patch adds a level of indirection to the deployment fixture -
instead of using the fixed 'ansible_all' handle, we're defining
a new one - 'ansible_vms_to_deploy'. The new fixture takes a default
value of 'ansible_all', but it's possible to override this for other
suites. This is very benefitial for the HE suite since with 'ansible_all'
we'd be trying to address the HE VM which doesn't exist at the moment
of deployment.

The real fix for this problem will be removing the HE VM's definition
from LagoInitfile and switching to a dynamic ansible inventory. Work
on these two tasks is already in progress.

Change-Id: I225e95f224d9f0e0110c8a2ee2e3feb2ca84bdc0
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>