he: Workaround for broken 'hostname' package
Recent 'hostname' package version introduced a dependency cycle
in systemd's units [1]. One of the side effects of this issue is that
the 'NetworkManager' doesn't start during system boot randomly.

We've implemented a workaround for this for ost-images themselves [2],
but that didn't cover the HE appliance, which needs its own fix.

[1] https://bugzilla.redhat.com/1959720
[2] https://gerrit.ovirt.org/#/c/ost-images/+/114997/

Change-Id: I61e809f8ac985b5aaf41e5bb5c1e179ec7aa0b96
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>