sd_reattach.py: consider OVF update before deactivating SD
The test for deactivating the storage domain may fail while an OVF
update is being occurred, with the following error:
Storage Domain OVF data is currently being updated.

As in case of running tasks, the test tries to deactivate the
storage domain again (in intervals of 10 seconds).
In case of failure in deactivating the storage domain after 3 minutes,
the test should fail.

Change-Id: I40bd2d9c3b11941bbf5a322e99ddae476d3142c4
Signed-off-by: Shani Leviim <sleviim@redhat.com>