network: test qos - ip configuration unnecessary
Assigning an IP is unnecessary while attaching a network to a NIC on the
host. This is because the goal of this test is to verify that the QoS
profiles are attached to the VM's VNIC when it is connected to this
network.
This attachment action is decoupled from IP assignment because QoS (aka
traffic control) is a layer 2 configuration while IP assignment is done
on layer 3.

Change-Id: I05dd04183750159cb1f264a412643ceeb4efdb9e
Signed-off-by: Eitan Raviv <eraviv@redhat.com>