deployment: Disable broken libvirt on hosts
We have at least 2 issues with new libvirt 7.4 - one is a problem
with selinux policies [1] and the other is a regression
in 'virDomainAttachDevice' [2].

To work around this problem we can downgrade the broken libvirt
version. With that we also need to downgrade qemu-kvm, cause the new
one causes problems with older libvirt.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=1966842
[2] https://bugzilla.redhat.com/show_bug.cgi?id=1970277

Change-Id: I43a02e4684ab11c79c528b163b953890d59291db
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>