network: host post install - rerun wait for up
After host installation is complete the host may move in and out of 'up'
status until InitVdsOnUpCommand completes and the host status
stabilizes. Waiting for the first time the host reaches 'up' status is
therefore not enough as a signal that installation is finished.
This patch re-runs host wait for up status, taking a guess that if after
six reruns with 10 sec intervals between them the host is up it means
that the transient phase is over.

Change-Id: Ie7bbe8069a7a9dde9c019312b8290657b0f5b3f8
Signed-off-by: Eitan Raviv <eraviv@redhat.com>