network: test required net - ip configuration unnecessary
Assigning an IP is unnecessary when testing the host status due to
changes in the network 'required' field on the cluster.
This is because the 'required' configuration affects the state of the
host only in engine, and is not affected by or has any effect on the
host machine itself.

Change-Id: Iab99879a3ff984331768250ce0a4b50673e8340b
Signed-off-by: Eitan Raviv <eraviv@redhat.com>