package_mgmt: Simplify check for unused repos
The output of 'dnf repo-pkgs' command is breaking lines from time
to time, i.e.:

 Installed Packages
 ovirt-ansible-collection.noarch 1.5.1-0.1.master.20210614081338.el8 @extra-src-1
 ovirt-imageio-common.x86_64     2.2.0-0.202106141241.gitd62c309.el8 @extra-src-1
 ovirt-imageio-daemon.x86_64     2.2.0-0.202106141241.gitd62c309.el8 @extra-src-1
 python3-ovirt-engine-sdk4.x86_64
                                 4.4.13-1.20210614.git6df453a.el8 @extra-src-1

This causes failures when trying to create 'Package' namedtuples.
We had the 'stty cols 300' to help with this problem, but it seems
it doesn't work all the time. Most probably dnf itself miscalculates
the spacing between the strings.

Since now we report oVirt package versions in
'report_ovirt_packages_versions' we don't need to create 'Package'
namedtuples from the output of 'dnf repo-pkgs' anymore.
A simple check on the amount of lines reported does the trick as well.

Change-Id: I9a8f9b76411dbbb4cd5e9b4aebfc5185fd00a95b
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>