ost_utils: collect and cache all ansible facts at once
Collecting all the facts at once is cheaper because it generates
one ansible request only instead of one per fact.
It also enables passing only a dict of facts to consuming layers,
freeing them from depending on the constructs and specific objects
of the module_mappers layer. A python dict is a universal construct
that does not create coupling between producer and consumer.

Signed-off-by: Eitan Raviv <eraviv@redhat.com>
Change-Id: Ie52e19777b7d6c5d9ab55a84b2273cb3a380e5e2