network: consume ansible facts as api from ost_utils
Depending on internal ost_utils constructs for ansible facts is fragile
and error prone. Instead, consuming a dictionary of facts is universal,
and dependent only on the ansible api which is what facts consumers
should be and are accountable for.
Also, all facts are cached in the object for usage throughtout the session
without the need to make repetitive ansible requests to the machines
for each property. This should prove more time efficient.

Signed-off-by: Eitan Raviv <eraviv@redhat.com>
Change-Id: I0112d03dc183e77fc3303007a0c9d7fe9606de58