UI: Change the navigation in the left menu
The navigation in the left menu has caused intermittent problems
where the hover did not display the submenu and as a result the test
failed later when asserting the correct page is displayed.

Now the hover is attempted multiple times using selenium wait mechanism.
This should also reduce test execution time since it is not necessary
to use static timeouts any more.

Change-Id: Ie443e04e511c2afb8c25d18d8abbb14bf90bb0a2
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>