ost: ui tests video recording
There's a new beta feature in Selenium container images - the
possibility to run a sidekick container that will record
the test session with ffmpeg by attaching to the x server of
the container with the browser.
This would be useful to have since the screenshots we currently
use only cover the browsers' viewport and not the whole
virtual screen, so we're missing i.e. javascript alert popups.

Change-Id: I1a0f5be7d6ed9a1f4616d97936bd2d3b1d2a73a6
Signed-off-by: hbraha <hbraha@redhat.com>