backup: temporary fix OST backup test due to change in VM backup behavior
The VM backup behavior has changed:
1) New success & failure phases were added.
2) Backup is considered as finished not when the relevant DB table
entry disappears, but when the backup phase is changed to SUCCEEDED.
Commenting out an assert in "backup.py::perform_vm_backup()" method
as a temporary solution. The correct check should be to check for
"BackupPhase.SUCCEEDED" status vs. expecting an empty list.

Waiting till the up-to-date ovirtsdk4.types are used by the OST.
The follow-up patch is already ready and should be merged ASAP.

Signed-off-by: Pavel Bar <pbar@redhat.com>
Change-Id: I4d0fdf0f5bf220ef77984b4081d70bc581586946