ost_utils: Use vdsm coverage reports functionality
This patch glues together the already implemented pieces that allow
us to generate vdsm coverage in OST runs.

Given that we've recently switched to FIPS mode in all OST runs
and the fact that the 'python3-coverage' package uses md5
internally [1] we need a small hack to make it use sha1 instead.
Otherwise the report generation ends with an error:

 ValueError: [digital envelope routines: EVP_DigestInit_ex] disabled for FIPS

[1] https://github.com/nedbat/coveragepy/blob/4e75bc6baa20e0c4a9377ade6310039e47d61897/coverage/misc.py#L203

Change-Id: Ic80715dc947faa23a85ec282bee972966091ce32
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>