host_utils: Don't fail on poking hosts
Since recent introduction of refactoring in host status area [1],
we've had one failure in CI, where an attempt to activate a host
ended with a lock error because a different operation was in progress.
The goal of the logic is to wait until the hosts are in working
state, so we should not fail in this case.

This patch adds exception handling that ignores the the error
in the scenario described above.

[1] commit hash ab24d2d, change id If4953a552bd

Change-Id: I49efaeb89f0abe136e76791621679879de957ec7
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>