network: test sync all hosts - ip configuration unnecessary
Assigning an IP when testing network sync status is unnecessary because
it does not affect this status in this test. Therefore it is not related
to the tested behavior.
The IP configuration is removed so that the test is agnostic to the
selected address family (ipv4 or ipv6) used in the test session.
Change-Id: I5cdf64fe8638da81cbed40d707987b07c198bf77
Signed-off-by: Eitan Raviv <eraviv@redhat.com>