basic: external network might have been imported
The test adds a network to the provider and then goes on to import it
into engine. But between these two steps engine may run its periodic
SyncNetworkProviderCommand which imports the network into engine causing
the import instantiated by the test to fail because the network already
exists. This patch anticipates the error and allows the test to continue
if it happens.
The test is also extended and tries to re-import the external network
intentionally to check that a re-import is impossible and guard against
regressions in this flow.
Change-Id: Iae9a7fc3e7215316f3a8c615c036663bac900c5b
Signed-off-by: Eitan Raviv <eraviv@redhat.com>