deactivate_sd: wait for FINISHED status after OVF update
In case an OVF_UPDATE is being executed on the storage domain while
deactivate it, the first iteration fails while the test keeps
running and waits for the second iteration to get finished successfully.

After it was succeeded, the tasks' statuses set looks like
{failed, finished} so comparing it to {finished} returns false and fails
the test.

This patch changes the validation to look for a finished status in the
returned set instead.

Change-Id: Iadbd1e02abef1a7bc31ea18dae45dc8157e3e0dd
Signed-off-by: Shani Leviim <sleviim@redhat.com>