backend: Inspect OST environment with virsh
The virsh-based backend relies on the output of 'ost_status'
function from 'lagofy.sh'. This is actually a reminescence
of what we've been doing with lago - parsing the output
of 'lago status' to extract information about the OST env.
Instead of doing that we should query libvirt with virsh
directly instead.

This patch introduces 3 namedtuples populated with such
queries - 'VMInfo', 'NICInfo' and 'NetworkInfo'. They will
be used in subsequent patches to reimplement the existing
functions.

Change-Id: Iff51097afe3bf091b86f09d0f8e70e92c38ce413
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>