ost-images: Simplify using OST images without their rpm's
The rpm packages may not be installed but the images can still be
present.  For instance when ost-images directory is mounted to a
container, without the need to fetch the large packages into the
container or to reuse them in multiple containers.

While it is already possible to specify individual locations of images
and keys via numerous OST_IMAGES_* variables, it is cumbersome for
the purpose of using standard images and breaks in case anything in
that set of variables changes.  Let's introduce OST_IMAGES_DIR
variable specifying the directory containing OST images.

Change-Id: Ie717b7a26e2dde3facb43cb72b05aa8337ece1be
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>