backend: Add 'ip_mapping' interface
This patch introduces 'BaseBackend.ip_mapping' interface
and implementations for 'LagoBackend' and 'VirshBackend'.
The new interface is very similar to the existing
'BaseBackend.iface_mapping' - it maps host names to network names
to IP addresses.

Obtaining NIC names cannot work reliably as it's up to the guest
operating system to decide which naming scheme to use. We now
have access to IP addresses, so we should this instead.

With that, we 'BaseBackend.iface_mapping' is being deprecated.
We still have one test using it and it will be refactored
in subsequent patches.

Change-Id: I0ac1ee1417aed81d96ffdec479aedb3ba906a822
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>