he: drop adding host-0 IPv4 to /etc/hosts
Complicates IPv6 work and It shouldn't really be needed anyway. Except
it is then confusing setup code that switching deployment to IPv6 due to
the host-0 resolving its DNS to IPv6 by default. For now we can enforce
setup to use IPv4 only, once we supply the right IP information for HE
on IPv6 we can drop that

Change-Id: I8cc0b5f4280d77769cb09ab829eca5a18c3a7a14
Signed-off-by: Michal Skrivanek <michal.skrivanek@redhat.com>