podman: Make network backend configurable
When OST is run under root, slirp4netns is used as the network
backend.  But it can be necessary to specify this backend also in
other situations, e.g. when running OST in a Podman container.

Rather than handling the special case of running OST as root (which is
not recommended anyway), let's make the network backend configurable
via PODMAN_NETWORK_BACKEND environment variable.  And make it set in
the automation environment where OST is run as root.

Change-Id: I16b8d1edfa3039e70fb0f0003e1ab3508f0e6c63
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>