lagofy: generate DNS entries for all interfaces
Until now we've generated FQDNs for management interface only, but it is
useful to have names for e.g. storage network interfaces too so that we
can use FQDNs instead of hardcoded IPs in deploy scripts and tests. Now
we're generating FQDN for every interface on each host, and add that to
DNS server on management network.
The order of processing network definitions was changed so that the
management network is defined last and includes all IPs from other
networks.

Change-Id: I4b78a7b35df12af9d7875fa9a6d3e5fd9625e4de
Signed-off-by: Michal Skrivanek <michal.skrivanek@redhat.com>